requirements.txt
beautifulsoup4
docx
fitz
ffmpeg-python
openai
PyMuPDF
Pillow
PyQt5
python-docx
python-vlc
whisper
langchain
langchain_openai
langchain_anthropic
groq
langchain-community


